requirements/requirements.txt
# Core model & performance libs
transformers
accelerate
einops
tiktoken
protobuf
sentencepiece
bitsandbytes

# LANGCHAIN
langchain
langchain-community
langchain-huggingface

# FastAPI server
fastapi
uvicorn[standard]
jinja2

# RAG / vector search
datasets
sentence-transformers
faiss-gpu
nest_asyncio
playwright
html2text
googlesearch-python

# Utilities
numpy<2.0
wikipedia
